garmin_service/requirements.txt
garminconnect
fastapi
uvicorn[standard]
python-dotenv
